requirements.txt
fastapi
uvicorn
torch
torchvision
torchaudio
pillow
python-multipart
pandas
numpy
matplotlib
yolov5
